- fix compilation

+++ hi_scripting/scripting/api/ScriptingApi.cpp
@@ -5344,7 +5344,7 @@ void ScriptingApi::Synth::playNoteFromUI(int channel, int noteNumber, int veloci
 {
     CustomKeyboardState& state = getScriptProcessor()->getMainController_()->getKeyboardState();
     
-    state.injectMessage(MidiMessage::noteOn(channel, noteNumber, velocity));
+    state.injectMessage(MidiMessage::noteOn(channel, noteNumber, (uint8)velocity));
 }
 
 void ScriptingApi::Synth::noteOffFromUI(int channel, int noteNumber)
